--- src/pages/auth/find-id/FindId.tsx
@@ -1,3 +1,0 @@
-export default function FindId() {
-  return <div>FindId</div>;
-}
