+++ src/pages/auth/password/Password.tsx
