+++ src/routes/router.tsx
@@ -1,7 +1,6 @@
 import { createBrowserRouter, type RouteObject } from "react-router-dom";
-import FindId from "@/pages/auth/find-id/FindId";
 import LoginPage from "@/pages/auth/login/Login";
-import ResetPassword from "@/pages/auth/reset-password/ResetPassword";
+import ResetPassword from "@/pages/auth/password/Password";
 import SignUpPage from "@/pages/auth/signup/SignUp";
 import HomePage from "@/pages/home/Home";
 import RootLayout from "../layouts/RootLayout";
@@ -25,11 +24,6 @@ const routes: (RouteObject & { handle?: RouteHandle })[] = [
       {
         path: "signup",
         element: <SignUpPage />,
-        handle: { bgColor: "bg-white", hideBottomNav: true },
-      },
-      {
-        path: "find-id",
-        element: <FindId />,
         handle: { bgColor: "bg-white", hideBottomNav: true },
       },
       {
